feat: token is checked for expiration and refreshed as needed
